requirements.txt
Pillow
pillow-heif
dhash
click
sqlalchemy[asyncio]
aiosqlite
aiofiles
throttler
alembic
tqdm
whatimage
numba
filetype
black
loguru
perception
opencv-python
orjson